removing duplicate code in multiple spots.
