+++ hexrd/unitcell.py
@@ -2,6 +2,7 @@ import importlib.resources
 import numpy as np
 from hexrd import constants
 from hexrd import symmetry, symbols
+from hexrd.spacegroup import Allowed_HKLs
 from hexrd.ipfcolor import sphere_sector, colorspace
 import hexrd.resources
 import warnings
@@ -811,542 +812,6 @@ class unitcell:
         sth = self.wavelength * glen * 0.5
         return np.arcsin(sth)
 
-    def Allowed_HKLs(self, hkllist):
-        '''
-        this function checks if a particular g vector is allowed
-        by lattice centering, screw axis or glide plane
-        '''
-        hkllist = np.atleast_2d(hkllist)
-
-        centering = self.sg_hmsymbol[0]
-
-        if(centering == 'P'):
-            # all reflections are allowed
-            mask = np.ones([hkllist.shape[0], ], dtype=np.bool)
-
-        elif(centering == 'F'):
-            # same parity
-
-            seo = np.sum(np.mod(hkllist+100, 2), axis=1)
-            mask = np.logical_not(np.logical_or(seo == 1, seo == 2))
-
-        elif(centering == 'I'):
-            # sum is even
-            seo = np.mod(np.sum(hkllist, axis=1)+100, 2)
-            mask = (seo == 0)
-
-        elif(centering == 'A'):
-            # k+l is even
-            seo = np.mod(np.sum(hkllist[:, 1:3], axis=1)+100, 2)
-            mask = seo == 0
-
-        elif(centering == 'B'):
-            # h+l is even
-            seo = np.mod(hkllist[:, 0]+hkllist[:, 2]+100, 2)
-            mask = seo == 0
-
-        elif(centering == 'C'):
-            # h+k is even
-            seo = np.mod(hkllist[:, 0]+hkllist[:, 1]+100, 2)
-            mask = seo == 0
-
-        elif(centering == 'R'):
-            # -h+k+l is divisible by 3
-            seo = np.mod(-hkllist[:, 0]+hkllist[:, 1]+hkllist[:, 2]+90, 3)
-            mask = seo == 0
-        else:
-            raise RuntimeError(
-                'IsGAllowed: unknown lattice centering encountered.')
-
-        hkls = hkllist[mask, :]
-
-        if(not self.symmorphic):
-            hkls = self.NonSymmorphicAbsences(hkls)
-
-        return hkls.astype(np.int32)
-
-    def omitscrewaxisabsences(self, hkllist, ax, iax):
-        '''
-        this function encodes the table on pg 48 of
-        international table of crystallography vol A
-        the systematic absences due to different screw
-        axis is encoded here.
-        iax encodes the primary, secondary or tertiary axis
-        iax == 0 : primary
-        iax == 1 : secondary
-        iax == 2 : tertiary
-
-        @NOTE: only unique b axis in monoclinic systems
-        implemented as thats the standard setting
-        '''
-        if(self.latticeType == 'triclinic'):
-            '''
-                no systematic absences for the triclinic crystals
-            '''
-            pass
-
-        elif(self.latticeType == 'monoclinic'):
-
-            if(ax != '2_1'):
-                raise RuntimeError(
-
-                    'omitscrewaxisabsences: monoclinic\
-                     systems can only have 2_1 screw axis.')
-
-            '''
-                only unique b-axis will be encoded
-                it is the users responsibility to input
-                lattice parameters in the standard setting
-                with b-axis having the 2-fold symmetry
-            '''
-            if(iax == 1):
-                mask1 = np.logical_and(hkllist[:, 0] == 0, hkllist[:, 2] == 0)
-                mask2 = np.mod(hkllist[:, 1]+100, 2) != 0
-                mask = np.logical_not(np.logical_and(mask1, mask2))
-                hkllist = hkllist[mask, :]
-            else:
-                raise RuntimeError(
-                    'omitscrewaxisabsences: only\
-                     b-axis can have 2_1 screw axis')
-
-        elif(self.latticeType == 'orthorhombic'):
-
-            if(ax != '2_1'):
-                raise RuntimeError(
-                    'omitscrewaxisabsences: orthorhombic\
-                     systems can only have 2_1 screw axis.')
-
-            '''
-            2_1 screw on primary axis
-            h00 ; h = 2n
-            '''
-            if(iax == 0):
-                mask1 = np.logical_and(hkllist[:, 1] == 0, hkllist[:, 2] == 0)
-                mask2 = np.mod(hkllist[:, 0]+100, 2) != 0
-                mask = np.logical_not(np.logical_and(mask1, mask2))
-                hkllist = hkllist[mask, :]
-
-            elif(iax == 1):
-                mask1 = np.logical_and(hkllist[:, 0] == 0, hkllist[:, 2] == 0)
-                mask2 = np.mod(hkllist[:, 1]+100, 2) != 0
-                mask = np.logical_not(np.logical_and(mask1, mask2))
-                hkllist = hkllist[mask, :]
-
-            elif(iax == 2):
-                mask1 = np.logical_and(hkllist[:, 0] == 0, hkllist[:, 1] == 0)
-                mask2 = np.mod(hkllist[:, 2]+100, 2) != 0
-                mask = np.logical_not(np.logical_and(mask1, mask2))
-                hkllist = hkllist[mask, :]
-
-        elif(self.latticeType == 'tetragonal'):
-
-            if(iax == 0):
-                mask1 = np.logical_and(hkllist[:, 0] == 0, hkllist[:, 1] == 0)
-                if(ax == '4_2'):
-                    mask2 = np.mod(hkllist[:, 2]+100, 2) != 0
-                elif(ax in ['4_1', '4_3']):
-                    mask2 = np.mod(hkllist[:, 2]+100, 4) != 0
-                mask = np.logical_not(np.logical_and(mask1, mask2))
-                hkllist = hkllist[mask, :]
-
-            elif(iax == 1):
-                mask1 = np.logical_and(hkllist[:, 1] == 0, hkllist[:, 2] == 0)
-                mask2 = np.logical_and(hkllist[:, 0] == 0, hkllist[:, 2] == 0)
-                if(ax == '2_1'):
-                    mask3 = np.mod(hkllist[:, 0]+100, 2) != 0
-                    mask4 = np.mod(hkllist[:, 1]+100, 2) != 0
-
-                mask1 = np.logical_not(np.logical_and(mask1, mask3))
-                mask2 = np.logical_not(np.logical_and(mask2, mask4))
-
-                mask = ~np.logical_or(~mask1, ~mask2)
-                hkllist = hkllist[mask, :]
-
-        elif(self.latticeType == 'trigonal'):
-            mask1 = np.logical_and(hkllist[:, 0] == 0, hkllist[:, 1] == 0)
-
-            if(iax == 0):
-                if(ax in ['3_1', '3_2']):
-                    mask2 = np.mod(hkllist[:, 2]+90, 3) != 0
-            else:
-                raise RuntimeError(
-                    'omitscrewaxisabsences: trigonal\
-                     systems can only have screw axis on primary axis ')
-
-            mask = np.logical_not(np.logical_and(mask1, mask2))
-            hkllist = hkllist[mask, :]
-
-        elif(self.latticeType == 'hexagonal'):
-
-            mask1 = np.logical_and(hkllist[:, 0] == 0, hkllist[:, 1] == 0)
-
-            if(iax == 0):
-                if(ax == '6_3'):
-                    mask2 = np.mod(hkllist[:, 2]+100, 2) != 0
-
-                elif(ax in ['3_1', '3_2', '6_2', '6_4']):
-                    mask2 = np.mod(hkllist[:, 2]+90, 3) != 0
-
-                elif(ax in ['6_1', '6_5']):
-                    mask2 = np.mod(hkllist[:, 2]+120, 6) != 0
-            else:
-                raise RuntimeError(
-                    'omitscrewaxisabsences: hexagonal\
-                     systems can only have screw axis on primary axis ')
-
-            mask = np.logical_not(np.logical_and(mask1, mask2))
-            hkllist = hkllist[mask, :]
-
-        elif(self.latticeType == 'cubic'):
-            mask1 = np.logical_and(hkllist[:, 0] == 0, hkllist[:, 1] == 0)
-            mask2 = np.logical_and(hkllist[:, 0] == 0, hkllist[:, 2] == 0)
-            mask3 = np.logical_and(hkllist[:, 1] == 0, hkllist[:, 2] == 0)
-
-            if(ax in ['2_1', '4_2']):
-                mask4 = np.mod(hkllist[:, 2]+100, 2) != 0
-                mask5 = np.mod(hkllist[:, 1]+100, 2) != 0
-                mask6 = np.mod(hkllist[:, 0]+100, 2) != 0
-
-            elif(ax in ['4_1', '4_3']):
-                mask4 = np.mod(hkllist[:, 2]+100, 4) != 0
-                mask5 = np.mod(hkllist[:, 1]+100, 4) != 0
-                mask6 = np.mod(hkllist[:, 0]+100, 4) != 0
-
-            mask1 = np.logical_not(np.logical_and(mask1, mask4))
-            mask2 = np.logical_not(np.logical_and(mask2, mask5))
-            mask3 = np.logical_not(np.logical_and(mask3, mask6))
-
-            mask = ~np.logical_or(~mask1, np.logical_or(~mask2, ~mask3))
-
-            hkllist = hkllist[mask, :]
-
-        return hkllist.astype(np.int32)
-
-    def omitglideplaneabsences(self, hkllist, plane, ip):
-        '''
-        this function encodes the table on pg 47 of
-        international table of crystallography vol A
-        the systematic absences due to different glide
-        planes is encoded here.
-        ip encodes the primary, secondary or tertiary plane normal
-        ip == 0 : primary
-        ip == 1 : secondary
-        ip == 2 : tertiary
-
-
-        @NOTE: only unique b axis in monoclinic systems
-        implemented as thats the standard setting
-        '''
-        if(self.latticeType == 'triclinic'):
-            pass
-
-        elif(self.latticeType == 'monoclinic'):
-            if(ip == 1):
-                mask1 = hkllist[:, 1] == 0
-
-                if(plane == 'c'):
-                    mask2 = np.mod(hkllist[:, 2]+100, 2) != 0
-                elif(plane == 'a'):
-                    mask2 = np.mod(hkllist[:, 0]+100, 2) != 0
-                elif(plane == 'n'):
-                    mask2 = np.mod(hkllist[:, 0]+hkllist[:, 2]+100, 2) != 0
-
-                mask = np.logical_not(np.logical_and(mask1, mask2))
-                hkllist = hkllist[mask, :]
-
-        elif(self.latticeType == 'orthorhombic'):
-
-            if(ip == 0):
-                mask1 = hkllist[:, 0] == 0
-
-                if(plane == 'b'):
-                    mask2 = np.mod(hkllist[:, 1]+100, 2) != 0
-                elif(plane == 'c'):
-                    mask2 = np.mod(hkllist[:, 2]+100, 2) != 0
-                elif(plane == 'n'):
-                    mask2 = np.mod(hkllist[:, 1]+hkllist[:, 2]+100, 2) != 0
-                elif(plane == 'd'):
-                    mask2 = np.mod(hkllist[:, 1]+hkllist[:, 2]+100, 4) != 0
-
-                mask = np.logical_not(np.logical_and(mask1, mask2))
-                hkllist = hkllist[mask, :]
-
-            elif(ip == 1):
-                mask1 = hkllist[:, 1] == 0
-
-                if(plane == 'c'):
-                    mask2 = np.mod(hkllist[:, 2]+100, 2) != 0
-                elif(plane == 'a'):
-                    mask2 = np.mod(hkllist[:, 0]+100, 2) != 0
-                elif(plane == 'n'):
-                    mask2 = np.mod(hkllist[:, 0]+hkllist[:, 2]+100, 2) != 0
-                elif(plane == 'd'):
-                    mask2 = np.mod(hkllist[:, 0]+hkllist[:, 2]+100, 4) != 0
-
-                mask = np.logical_not(np.logical_and(mask1, mask2))
-                hkllist = hkllist[mask, :]
-
-            elif(ip == 2):
-                mask1 = hkllist[:, 2] == 0
-
-                if(plane == 'a'):
-                    mask2 = np.mod(hkllist[:, 0]+100, 2) != 0
-                elif(plane == 'b'):
-                    mask2 = np.mod(hkllist[:, 1]+100, 2) != 0
-                elif(plane == 'n'):
-                    mask2 = np.mod(hkllist[:, 0]+hkllist[:, 1]+100, 2) != 0
-                elif(plane == 'd'):
-                    mask2 = np.mod(hkllist[:, 0]+hkllist[:, 1]+100, 4) != 0
-
-                mask = np.logical_not(np.logical_and(mask1, mask2))
-                hkllist = hkllist[mask, :]
-
-        elif(self.latticeType == 'tetragonal'):
-
-            if(ip == 0):
-
-                mask1 = hkllist[:, 2] == 0
-
-                if(plane == 'a'):
-                    mask2 = np.mod(hkllist[:, 0]+100, 2) != 0
-
-                elif(plane == 'b'):
-                    mask2 = np.mod(hkllist[:, 1]+100, 2) != 0
-
-                elif(plane == 'n'):
-                    mask2 = np.mod(hkllist[:, 0]+hkllist[:, 1]+100, 2) != 0
-
-                elif(plane == 'd'):
-                    mask2 = np.mod(hkllist[:, 0]+hkllist[:, 1]+100, 4) != 0
-
-                mask = np.logical_not(np.logical_and(mask1, mask2))
-                hkllist = hkllist[mask, :]
-
-            elif(ip == 1):
-
-                mask1 = hkllist[:, 0] == 0
-                mask2 = hkllist[:, 1] == 0
-
-                if(plane in ['a', 'b']):
-                    mask3 = np.mod(hkllist[:, 1]+100, 2) != 0
-                    mask4 = np.mod(hkllist[:, 0]+100, 2) != 0
-
-                elif(plane == 'c'):
-                    mask3 = np.mod(hkllist[:, 2]+100, 2) != 0
-                    mask4 = mask3
-
-                elif(plane == 'n'):
-                    mask3 = np.mod(hkllist[:, 1]+hkllist[:, 2]+100, 2) != 0
-                    mask4 = np.mod(hkllist[:, 0]+hkllist[:, 2]+100, 2) != 0
-
-                elif(plane == 'd'):
-                    mask3 = np.mod(hkllist[:, 1]+hkllist[:, 2]+100, 4) != 0
-                    mask4 = np.mod(hkllist[:, 0]+hkllist[:, 2]+100, 4) != 0
-
-                mask1 = np.logical_not(np.logical_and(mask1, mask3))
-                mask2 = np.logical_not(np.logical_and(mask2, mask4))
-
-                mask = ~np.logical_or(~mask1, ~mask2)
-                hkllist = hkllist[mask, :]
-
-            elif(ip == 2):
-
-                mask1 = np.abs(hkllist[:, 0]) == np.abs(hkllist[:, 1])
-
-                if(plane in ['c', 'n']):
-                    mask2 = np.mod(hkllist[:, 2]+100, 2) != 0
-
-                elif(plane == 'd'):
-                    mask2 = np.mod(2*hkllist[:, 0]+hkllist[:, 2]+100, 4) != 0
-
-                mask = np.logical_not(np.logical_and(mask1, mask2))
-
-                hkllist = hkllist[mask, :]
-
-        elif(self.latticeType == 'trigonal'):
-
-            if(plane != 'c'):
-                raise RuntimeError(
-                    'omitglideplaneabsences: only\
-                     c-glide allowed for trigonal systems.')
-
-            if(ip == 1):
-
-                mask1 = hkllist[:, 0] == 0
-                mask2 = hkllist[:, 1] == 0
-                mask3 = hkllist[:, 0] == -hkllist[:, 1]
-                if(plane == 'c'):
-                    mask4 = np.mod(hkllist[:, 2]+100, 2) != 0
-                else:
-                    raise RuntimeError(
-                        'omitglideplaneabsences: only\
-                         c-glide allowed for trigonal systems.')
-
-            elif(ip == 2):
-                mask1 = hkllist[:, 1] == hkllist[:, 0]
-                mask2 = hkllist[:, 0] == -2*hkllist[:, 1]
-                mask3 = -2*hkllist[:, 0] == hkllist[:, 1]
-                if(plane == 'c'):
-                    mask4 = np.mod(hkllist[:, 2]+100, 2) != 0
-                else:
-                    raise RuntimeError(
-                        'omitglideplaneabsences: only\
-                         c-glide allowed for trigonal systems.')
-
-            mask1 = np.logical_and(mask1, mask4)
-            mask2 = np.logical_and(mask2, mask4)
-            mask3 = np.logical_and(mask3, mask4)
-
-            mask = np.logical_not(np.logical_or(
-                mask1, np.logical_or(mask2, mask3)))
-            hkllist = hkllist[mask, :]
-
-        elif(self.latticeType == 'hexagonal'):
-
-            if(plane != 'c'):
-                raise RuntimeError(
-                    'omitglideplaneabsences: only\
-                     c-glide allowed for hexagonal systems.')
-
-            if(ip == 2):
-                mask1 = hkllist[:, 0] == hkllist[:, 1]
-                mask2 = hkllist[:, 0] == -2*hkllist[:, 1]
-                mask3 = -2*hkllist[:, 0] == hkllist[:, 1]
-
-                mask4 = np.mod(hkllist[:, 2]+100, 2) != 0
-
-                mask1 = np.logical_and(mask1, mask4)
-                mask2 = np.logical_and(mask2, mask4)
-                mask3 = np.logical_and(mask3, mask4)
-
-                mask = np.logical_not(np.logical_or(
-                    mask1, np.logical_or(mask2, mask3)))
-
-            elif(ip == 1):
-                mask1 = hkllist[:, 1] == 0
-                mask2 = hkllist[:, 0] == 0
-                mask3 = hkllist[:, 1] == -hkllist[:, 0]
-
-                mask4 = np.mod(hkllist[:, 2]+100, 2) != 0
-
-            mask1 = np.logical_and(mask1, mask4)
-            mask2 = np.logical_and(mask2, mask4)
-            mask3 = np.logical_and(mask3, mask4)
-
-            mask = np.logical_not(np.logical_or(
-                mask1, np.logical_or(mask2, mask3)))
-            hkllist = hkllist[mask, :]
-
-        elif(self.latticeType == 'cubic'):
-
-            if(ip == 0):
-                mask1 = hkllist[:, 0] == 0
-                mask2 = hkllist[:, 1] == 0
-                mask3 = hkllist[:, 2] == 0
-
-                mask4 = np.mod(hkllist[:, 0]+100, 2) != 0
-                mask5 = np.mod(hkllist[:, 1]+100, 2) != 0
-                mask6 = np.mod(hkllist[:, 2]+100, 2) != 0
-
-                if(plane == 'a'):
-                    mask1 = np.logical_or(np.logical_and(
-                        mask1, mask5), np.logical_and(mask1, mask6))
-                    mask2 = np.logical_or(np.logical_and(
-                        mask2, mask4), np.logical_and(mask2, mask6))
-                    mask3 = np.logical_and(mask3, mask4)
-
-                    mask = np.logical_not(np.logical_or(
-                        mask1, np.logical_or(mask2, mask3)))
-
-                elif(plane == 'b'):
-                    mask1 = np.logical_and(mask1, mask5)
-                    mask3 = np.logical_and(mask3, mask5)
-                    mask = np.logical_not(np.logical_or(mask1, mask3))
-
-                elif(plane == 'c'):
-                    mask1 = np.logical_and(mask1, mask6)
-                    mask2 = np.logical_and(mask2, mask6)
-                    mask = np.logical_not(np.logical_or(mask1, mask2))
-
-                elif(plane == 'n'):
-                    mask4 = np.mod(hkllist[:, 1]+hkllist[:, 2]+100, 2) != 0
-                    mask5 = np.mod(hkllist[:, 0]+hkllist[:, 2]+100, 2) != 0
-                    mask6 = np.mod(hkllist[:, 0]+hkllist[:, 1]+100, 2) != 0
-
-                    mask1 = np.logical_not(np.logical_and(mask1, mask4))
-                    mask2 = np.logical_not(np.logical_and(mask2, mask5))
-                    mask3 = np.logical_not(np.logical_and(mask3, mask6))
-
-                    mask = ~np.logical_or(
-                        ~mask1, np.logical_or(~mask2, ~mask3))
-
-                elif(plane == 'd'):
-                    mask4 = np.mod(hkllist[:, 1]+hkllist[:, 2]+100, 4) != 0
-                    mask5 = np.mod(hkllist[:, 0]+hkllist[:, 2]+100, 4) != 0
-                    mask6 = np.mod(hkllist[:, 0]+hkllist[:, 1]+100, 4) != 0
-
-                    mask1 = np.logical_not(np.logical_and(mask1, mask4))
-                    mask2 = np.logical_not(np.logical_and(mask2, mask5))
-                    mask3 = np.logical_not(np.logical_and(mask3, mask6))
-
-                    mask = ~np.logical_or(
-                        ~mask1, np.logical_or(~mask2, ~mask3))
-                else:
-                    raise RuntimeError(
-                        'omitglideplaneabsences: unknown\
-                         glide plane encountered.')
-
-                hkllist = hkllist[mask, :]
-
-            if(ip == 2):
-
-                mask1 = np.abs(hkllist[:, 0]) == np.abs(hkllist[:, 1])
-                mask2 = np.abs(hkllist[:, 1]) == np.abs(hkllist[:, 2])
-                mask3 = np.abs(hkllist[:, 0]) == np.abs(hkllist[:, 2])
-
-                if(plane in ['a', 'b', 'c', 'n']):
-                    mask4 = np.mod(hkllist[:, 2]+100, 2) != 0
-                    mask5 = np.mod(hkllist[:, 0]+100, 2) != 0
-                    mask6 = np.mod(hkllist[:, 1]+100, 2) != 0
-
-                elif(plane == 'd'):
-                    mask4 = np.mod(2*hkllist[:, 0]+hkllist[:, 2]+100, 4) != 0
-                    mask5 = np.mod(hkllist[:, 0]+2*hkllist[:, 1]+100, 4) != 0
-                    mask6 = np.mod(2*hkllist[:, 0]+hkllist[:, 1]+100, 4) != 0
-
-                else:
-                    raise RuntimeError(
-                        'omitglideplaneabsences: unknown\
-                         glide plane encountered.')
-
-                mask1 = np.logical_not(np.logical_and(mask1, mask4))
-                mask2 = np.logical_not(np.logical_and(mask2, mask5))
-                mask3 = np.logical_not(np.logical_and(mask3, mask6))
-
-                mask = ~np.logical_or(~mask1, np.logical_or(~mask2, ~mask3))
-
-                hkllist = hkllist[mask, :]
-
-        return hkllist
-
-    def NonSymmorphicAbsences(self, hkllist):
-        '''
-        this function prunes hkl list for the screw axis and glide
-        plane absences
-        '''
-        planes = constants.SYS_AB[self.sgnum][0]
-
-        for ip, p in enumerate(planes):
-            if(p != ''):
-                hkllist = self.omitglideplaneabsences(hkllist, p, ip)
-
-        axes = constants.SYS_AB[self.sgnum][1]
-        for iax, ax in enumerate(axes):
-            if(ax != ''):
-                hkllist = self.omitscrewaxisabsences(hkllist, ax, iax)
-
-        return hkllist
-
     def ChooseSymmetric(self, hkllist, InversionSymmetry=True):
         '''
         this function takes a list of hkl vectors and
@@ -1426,7 +891,7 @@ class unitcell:
                             for ik in np.arange(kmax, kmin, -1)
                             for il in np.arange(lmax, lmin, -1)])
 
-        hkl_allowed = self.Allowed_HKLs(hkllist)
+        hkl_allowed = Allowed_HKLs(self.sgnum, hkllist)
 
         hkl = []
         dsp = []
